Possible fix for the #11
It seems that if you have a short lived XML document and perform an path query on the document then there's a chance (based on order of release in autorelease pool) that the xpath nodes will get freed before the document node is freed.

This causes a crash.

This fixes the issue by putting the nodes from the xpath query into the document's node pool and then explicitly freeing (via a new invalidate method) the nodes in the node pool before the document is dealloced.